pyproject.toml
[project]
name = "uv-people-screening"
version = "0.1.0"
description = "People screening application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.0",
    "pandas-stubs>=2.2.3.250527",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "weasyprint>=65.1",
]

[dependency-groups]
dev = [
    "uv-people-screening",
]

[tool.uv.sources]
uv-people-screening = { workspace = true }
